cli not write if no change & react improve example interactivity

+++ packages/gqless-cli/src/writeGenerate.ts
@@ -15,6 +15,14 @@ export async function writeGenerate(
 
   await mkdirp(dirname(destinationPath));
 
+  if (fs.existsSync(destinationPath)) {
+    const existingCode = await fs.promises.readFile(destinationPath, {
+      encoding: 'utf-8',
+    });
+
+    if (existingCode === code) return destinationPath;
+  }
+
   await fs.promises.writeFile(destinationPath, code, {
     encoding: 'utf-8',
   });
